# swagger/swaggerTrnx.yaml
openapi: 3.0.0
info:
  title: Transaction API
  version: 1.0.0
  description: API for managing transactions, user balances, admin deposits, and Stripe bank cash-in.
servers:
  - url: https://dmoney.roadtocareer.net
    description: Production server
  - url: http://localhost:5000
    description: Local server
paths:
...
  /transaction/deposit:
    post:
      summary: Deposit money to an account (Agent → Customer)
      description: |
        Allows an authenticated user with the role of **Agent** to deposit money into a Customer account.
        - Amount must be between the configured minimum and maximum (default 10–10,000 tk).
        - A commission is earned by the Agent.
        - The token owner must be the `from_account` holder.
      operationId: deposit
      tags:
        - Transactions
      parameters:
        - $ref: '#/components/parameters/XAuthSecretKeyParam'
      security:
        - BearerAuth: []
        - XAuthSecretKey: []
      requestBody:
        description: Data for depositing money to an account.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from_account
                - to_account
                - amount
              properties:
                from_account:
                  type: string
                  example: "01700000001"
                  description: Phone number of the Agent account making the deposit.
                to_account:
                  type: string
                  example: "01700000002"
                  description: Phone number of the Customer account receiving the deposit.
                amount:
                  type: number
                  example: 500
                  description: Amount to deposit (must be between 10 and 10,000 tk).
      responses:
        '201':
          description: Deposit was successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Deposit successful
                  trnxId:
                    type: string
                    example: "TXN123456"
                  commission:
                    type: number
                    example: 12.5
                  currentBalance:
                    type: number
                    example: 487.5
        '208':
          description: Non-successful conditions (insufficient balance, role mismatch, limit exceeded, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Insufficient balance
                  currentBalance:
                    type: number
                    example: 300.0
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Account does not exist
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /user/login or /user/verify-otp

    XAuthSecretKey:
      type: apiKey
      in: header
      name: X-AUTH-SECRET-KEY
      description: Partner/admin secret key required for transaction API authentication
  parameters:
    XAuthSecretKeyParam:
      name: X-AUTH-SECRET-KEY
      in: header
      description: Partner key required for transaction API authentication
      required: true
      schema:
        type: string